requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
defusedxml==0.7.1
dj-database-url==3.0.0
Django==5.2.2
django-cors-headers==4.7.0
django-extensions==4.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1   #here
dotenv==0.9.9
flake8==7.3.0
gunicorn==23.0.0
idna==3.10
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.9.0
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
setuptools==78.1.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.9.0



#djangorestframework-authtoken==26.0.1
djoser
coverage

dj-database-url
